requirements.txt
roboflow==1.1.37
python-dotenv==1.0.1
ultralytics==8.2.79